+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 ---------
+## 1.2.0
+##### Domain
+* Dynamic loading for shard constants
+* Removed: Shard.baseUrl
+
 ## 1.1.0
 ##### Domain
 * Bugfix: Fixed wrong REST Api URLs